# Optimize string.join for IList<string> with char separator