@stephentoub 2021-08-04
> The below benchmark includes StringJoinIntList to verify that performance doesn't degrade for joining lists of non-string types

The JIT will optimize the check away for value types.  To validate impact here, you'd need to test it against some other reference type.