@stephentoub 2021-08-05
(For anyone following along, this is based on my comment: https://github.com/dotnet/runtime/pull/44032#discussion_r682234581)