@johnthcall 2021-08-05
@stephentoub I've replaced benchmark StringJoinIntList  with StringJoinClassList and updated the description.