@johnthcall 2021-08-06
Another option would be to add a new overload of Join(char, IEnumerable<String>) to match the Join(String, IEnumerable<String>) method. This would make the type check of T not necessary but would require an an extra JoinCore and the method calls to it to align with the IEnumerable<T>