@stephentoub 2021-08-07
> Another option

We could consider adding an overload,  but it's too late to do that for .NET 6.