@danmoseley 2021-08-11
Thanks for another perf contribution @johnthcall. We welcome them!